+++ benches/h3/latlng_to_cell.rs
@@ -4,7 +4,7 @@ use h3o::{LatLng, Resolution};
 pub fn bench(c: &mut Criterion) {
     let mut group = c.benchmark_group("latLngToCell");
 
-    let ll = LatLng::new(0.8526733842426371, 0.04141688057720644).expect("hex");
+    let ll = LatLng::new(48.85458622023985, 2.373012457671282).expect("hex");
     for resolution in 0..=15 {
         group.bench_with_input(
             BenchmarkId::new("h3o/Hexagon", resolution),
@@ -45,8 +45,8 @@ fn bench_h3o(b: &mut Bencher<'_>, ll: LatLng, resolution: u8) {
 fn bench_h3(b: &mut Bencher<'_>, ll: LatLng, resolution: u8) {
     let mut out: u64 = 0;
     let ll = h3ron_h3_sys::LatLng {
-        lat: ll.lat(),
-        lng: ll.lng(),
+        lat: ll.lat_radians(),
+        lng: ll.lng_radians(),
     };
     b.iter(|| unsafe {
         h3ron_h3_sys::latLngToCell(